# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Space Reservation API",
        "description": "API REST para sistema de reserva de espacios para eventos. Permite a los usuarios explorar espacios disponibles, hacer reservas y gestionarlas.",
        "contact": {
            "name": "Soporte API",
            "email": "admin@espacios.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor de desarrollo"
        }
    ],
    "paths": {
...
        "/reservations": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Listar reservaciones del usuario autenticado",
                "operationId": "c0ead7ee8da750a341eeff509f3af8b9",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrar por estado",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "confirmed",
                                "cancelled",
                                "pending"
                            ]
                        }
                    },
                    {
                        "name": "from_date",
                        "in": "query",
                        "description": "Desde fecha (Y-m-d)",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to_date",
                        "in": "query",
                        "description": "Hasta fecha (Y-m-d)",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de reservaciones",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Reservation"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Crear nueva reservación",
                "operationId": "ab4d7952bb3ef1ccd7724e8d862b034b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "space_id",
                                    "start_time",
                                    "end_time"
                                ],
                                "properties": {
                                    "space_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "format": "datetime",
                                        "example": "2025-01-20 09:00:00"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "format": "datetime",
                                        "example": "2025-01-20 12:00:00"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Reunión de equipo de desarrollo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reservación creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Reservación creada exitosamente"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Reservation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "409": {
                        "description": "Conflicto - El espacio no está disponible"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/reservations/{id}": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Obtener detalle de una reservación",
                "operationId": "49499ada3f4d3680e147782d33573252",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la reservación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle de la reservación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reservation"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Reservación no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Actualizar reservación",
                "operationId": "4a219e5cfd9760d12da084929201d042",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la reservación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "start_time": {
                                        "type": "string",
                                        "format": "datetime"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "format": "datetime"
                                    },
                                    "notes": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "confirmed",
                                            "cancelled",
                                            "pending"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reservación actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Reservación actualizada exitosamente"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Reservation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Reservación no encontrada"
                    },
                    "409": {
                        "description": "Conflicto de horario"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Cancelar/eliminar reservación",
                "operationId": "b97e0df7036fa97413d6b819cf61a11e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la reservación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reservación eliminada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Reservación eliminada exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Reservación no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/reservations/{id}/cancel": {
            "patch": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Cancelar una reservación (cambiar estado)",
                "operationId": "70237665fb6d7ef14914c29ac4190d5d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la reservación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reservación cancelada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Reservación cancelada exitosamente"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Reservation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Reservación no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Juan Pérez"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "juan@ejemplo.com"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "admin",
                            "user"
                        ],
                        "example": "user"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "datetime",
                        "example": "2025-01-01T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "datetime",
                        "example": "2025-01-01T00:00:00.000000Z"
                    }
                },
                "type": "object"
            },
            "Space": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Sala de Conferencias Principal"
                    },
                    "description": {
                        "type": "string",
                        "example": "Sala amplia con capacidad para 30 personas"
                    },
                    "capacity": {
                        "type": "integer",
                        "example": 30
                    },
                    "location": {
                        "type": "string",
                        "example": "Edificio A - Piso 2"
                    },
                    "amenities": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "WiFi",
                            "Proyector",
                            "Pizarra",
                            "Aire Acondicionado"
                        ]
                    },
                    "image_url": {
                        "type": "string",
                        "example": "https://example.com/sala.jpg",
                        "nullable": true
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "datetime"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "datetime"
                    }
                },
                "type": "object"
            },
            "Reservation": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "space_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "start_time": {
                        "type": "string",
                        "format": "datetime",
                        "example": "2025-01-20 09:00:00"
                    },
                    "end_time": {
                        "type": "string",
                        "format": "datetime",
                        "example": "2025-01-20 12:00:00"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "confirmed",
                            "cancelled",
                            "pending"
                        ],
                        "example": "confirmed"
                    },
                    "notes": {
                        "type": "string",
                        "example": "Reunión de equipo",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "datetime"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "datetime"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "space": {
                        "$ref": "#/components/schemas/Space"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingrese el token de autenticación",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}